Adding javadocs, extracting StreamBuilder class
